Start using "takari." prefix for control properties
There are several properties in our plugin definitions that can be made
more convenient to change by providing a control property. It's much
easier to change a property in your POM than redefining the plugin
block to override a property. There are many more that we can
control and as we move forward we will prefix them as not
to collide with other properties that may have the same
name but used in other plugins.